feat(core): export the comprehension static extractor + compile driver
Promote the compiled-comprehension substrate's static extractor and diff-
scoped run driver from the CLI into @harness-engineering/core's public
surface, so library consumers (e.g. pnyon's per-Outpost comprehension)
drive the substrate through the same IO-injected orchestration the
`harness comprehend` CLI uses — without depending on the CLI package.

Moved to core/comprehension:
- static-extractor.ts: createStaticExtractor + render helpers over core's
  TypeScriptParser
- run.ts: runComprehend / runComprehendCheck / runComprehendStats /
  mapWithConcurrency + Comprehend* types
- reentrancy.ts: withComprehensionActive / isComprehensionReentrant /
  REENTRANCY_ENV (pure env guard)

The CLI modules now re-export the core implementation for back-compat;
every existing importer keeps working. No behaviour change. core:minor,
cli:patch (changeset included). core 80 + cli 180 comprehension tests
green; typecheck clean.
