+++ packages/core/tests/comprehension/static-extractor.test.ts
